Rename compact_cow_str.rs to cow_rc_str.rs
