+++ src/lib.rs
@@ -91,7 +91,7 @@ pub use nth::parse_nth;
 pub use serializer::{ToCss, CssStringWriter, serialize_identifier, serialize_string, TokenSerializationType};
 pub use parser::{Parser, Delimiter, Delimiters, SourcePosition, ParseError, BasicParseError, ParserInput};
 pub use unicode_range::UnicodeRange;
-pub use compact_cow_str::CowRcStr;
+pub use cow_rc_str::CowRcStr;
 
 // For macros
 #[doc(hidden)] pub use macros::_internal__to_lowercase;
@@ -117,7 +117,7 @@ mod color;
 mod nth;
 mod serializer;
 mod unicode_range;
-mod compact_cow_str;
+mod cow_rc_str;
 
 #[cfg(test)] mod tests;
 #[cfg(test)] mod size_of_tests;
