+++ src/tokenizer.rs
@@ -10,7 +10,7 @@ use std::char;
 use std::ascii::AsciiExt;
 use std::i32;
 
-use compact_cow_str::CowRcStr;
+use cow_rc_str::CowRcStr;
 use self::Token::*;
